Q&A_Chatbot/requirements.txt
langchain-openai
pandas
numpy
python-dotenv
langchain-community
langchain
langchain-core
streamlit